Add ETCD to mgmt cluster types
