+++ pkg/apis/rke.cattle.io/v1/etcdsnapshot_types.go
@@ -1,6 +1,7 @@
 package v1
 
 import (
+	mgmt "github.com/rancher/rancher/pkg/apis/management.cattle.io/v3"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
@@ -13,72 +14,7 @@ const (
 )
 
 // ETCDSnapshotS3 defines S3 snapshot configuration for ETCD backups.
-type ETCDSnapshotS3 struct {
-	// Endpoint is the S3 endpoint used for snapshot operations.
-	// If this field is not explicitly set, the 'defaultEndpoint' value from the referenced CloudCredential will be used.
-	// +nullable
-	// +optional
-	Endpoint string `json:"endpoint,omitempty"`
-
-	// EndpointCA is the CA certificate for validating the S3 endpoint.
-	// This can be either a file path (e.g., "/etc/ssl/certs/my-ca.crt")
-	// or the CA certificate content, in base64-encoded or plain PEM format.
-	// If this field is not explicitly set, the 'defaultEndpointCA' value from the referenced CloudCredential will be used.
-	// +nullable
-	// +optional
-	EndpointCA string `json:"endpointCA,omitempty"`
-
-	// SkipSSLVerify defines whether TLS certificate verification is disabled.
-	// If this field is not explicitly set, the 'defaultSkipSSLVerify' value
-	// from the referenced CloudCredential will be used.
-	// +optional
-	SkipSSLVerify bool `json:"skipSSLVerify,omitempty"`
-
-	// Bucket is the name of the S3 bucket used for snapshot operations.
-	// If this field is not explicitly set, the 'defaultBucket' value from the referenced CloudCredential will be used.
-	// An empty bucket name will cause a 'failed to initialize S3 client: s3 bucket name was not set' error.
-	// +kubebuilder:validation:MaxLength=63
-	// +nullable
-	// +optional
-	Bucket string `json:"bucket,omitempty"`
-
-	// Region is the S3 region used for snapshot operations. (e.g., "us-east-1").
-	// If this field is not explicitly set, the 'defaultRegion' value from the referenced CloudCredential will be used.
-	// +nullable
-	// +optional
-	Region string `json:"region,omitempty"`
-
-	// CloudCredentialName is the name of the secret containing the
-	// credentials used to access the S3 bucket.
-	// The secret is expected to have the following keys:
-	// - accessKey [required]
-	// - secretKey [required]
-	// - defaultRegion
-	// - defaultEndpoint
-	// - defaultEndpointCA
-	// - defaultSkipSSLVerify
-	// - defaultBucket
-	// - defaultFolder
-	// Fields set directly in this spec (`ETCDSnapshotS3`) take precedence over the corresponding
-	// values from the CloudCredential secret. This field must be in the format of "namespace:name".
-	// +nullable
-	// +optional
-	CloudCredentialName string `json:"cloudCredentialName,omitempty"`
-
-	// Retention defines the number of snapshots to retain in the S3 bucket.
-	// Older snapshots beyond this retention count will be deleted.
-	// If this field is not explicitly set, the retention value from the etcd retention will be used.
-	// +optional
-	// +nullable
-	// +kubebuilder:validation:Minimum=0
-	Retention int `json:"retention,omitempty"`
-
-	// Folder is the name of the S3 folder used for snapshot operations.
-	// If this field is not explicitly set, the folder from the referenced CloudCredential will be used.
-	// +nullable
-	// +optional
-	Folder string `json:"folder,omitempty"`
-}
+type ETCDSnapshotS3 = mgmt.ETCDSnapshotS3
 
 // +genclient
 // +kubebuilder:resource:path=etcdsnapshots,scope=Namespaced
